Update rw_skeleton.c
For some reason ./rw 1 will actually pass the getopt without triggering the ? case (which it should). This command line parsing should work out of the box with this but it isn't. I'm going to sleep on it and see if I can see the fix in the morning. 

TODO: Still need to figure out the seg fault when passing in the unsigned int into the srand function call.

Otherwise I think this is fairly good to go.