Update rw_skeleton.c
Removed some debug. The threads all seem to run successfully. The creation of the read needs to be addressed.

In reader, rest() is run READER_ITR * 100 where READER_ITR = 100000 for the test case.
In writer, rest() is run WRITE_ITR * 1 where WRITE_ITR = 100000 for the test case.

As you can see there is orders of magnitude difference in the implementation and a significant lag in the creation of the read logs and the subsequent iterations relative to that of write.

I'm not certain this SHOULD be the way it works. This needs to be investigated.